Poll for load in CSP frame-ancestors tests
These tests are all timing out in Firefox because we don't send either
a load or an error event for CSP-blocked loads. To work around this,
poll the iframe for the load, assuming it's complete when we see a
non-about:blank document with readyState complete (or an exception
from trying to access a cross-origin resource).

Differential Revision: https://phabricator.services.mozilla.com/D62447

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1614674
gecko-commit: a2bf92004940994b68132da4d82c3965af991535
gecko-integration-branch: autoland
gecko-reviewers: bzbarsky