Call SetNeedsResizeObserverUpdate in LayoutBox
Currently, a resize observer update is only carried out upon a change in
an element's border box. In order to observe changes to the content box,
we call SetNeedsResizeObserverUpdate in LayoutBox::StyleDidChange.
The change is implemented in this location in particular due to the
availability of old_style and new_style, which allows for the comparison
of content boxes and writing mode. Alternative locations such as
LayoutObject::SetNeedsOverflowRecalc (where
SetShouldCheckForPaintInvalidation is called) and
LayoutObject::SetNeedsLayout were explored as alternatives but did not
seem to provide the right context to compare old and new styles (based
on my investigation).

With this change, all WPT observe.html tests now pass.

Bug: 1034329
Change-Id: Ie2225ad2865ee15d82be548c1455949946734074